requirements.txt
dnspython==2.6.1
pymongo==3.13.0
python-dotenv==1.0.1
apache-airflow==2.8.4
apache-airflow-providers-dbt-cloud==3.1.0
apache-airflow-providers-mongo==3.1.1
apache-airflow-providers-google==10.21.1
apache-airflow-providers-amazon==8.19.0
# apache-airflow-providers-celery==3.7.3
importlib-metadata==6.11.0


# apache-airflow==2.10.0
# apache-airflow-providers-amazon==8.19.0
# apache-airflow-providers-common-compat==1.1.0
# apache-airflow-providers-common-io==1.4.0
# apache-airflow-providers-common-sql==1.15.0
# apache-airflow-providers-dbt-cloud==3.1.0
# apache-airflow-providers-fab==1.2.2
# apache-airflow-providers-ftp==3.10.1
# apache-airflow-providers-google==10.21.1
# apache-airflow-providers-http==4.12.0
# apache-airflow-providers-imap==3.6.1
# apache-airflow-providers-mongo==3.1.1
# apache-airflow-providers-smtp==1.7.1
# apache-airflow-providers-sqlite==3.8.2
